No more warnings in code. Removed System::Proxy
Proxy was removed because I got troubles with destructor call

+++ source/math/random.cpp
@@ -51,6 +51,6 @@ namespace Math
 	float Random::Noise(int x)
 	{
 		x = ( x << 13 ) ^ x;
-		return ( 1.0 - ( (x * (x * x * 15731 + 789221) + 1376312589) & 0x7fffffff) / 1073741824.0);
+		return float( 1.0 - ( (x * (x * x * 15731 + 789221) + 1376312589) & 0x7fffffff) / 1073741824.0);
 	}
 }
